python/requirements.txt
pycryptodomex >= 3.6
secp256k1 >= 0.14.0
